Convert values to floats
The numpy float32 types aren't JSON serializable. For some reason I have
to explicitly cast the values computed for scalars to a float, but I
didn't have to do this for any of the calculations in the vector
variables.

+++ services/api/src/unified_graphics/diag.py
@@ -69,11 +69,15 @@ class ScalarDiag:
         counts, bin_edges = np.histogram(data, bins="auto")
 
         bins = [
-            Bin(lower=bin_edges[i], upper=bin_edges[i + 1], value=value)
+            Bin(
+                lower=float(bin_edges[i]),
+                upper=float(bin_edges[i + 1]),
+                value=float(value),
+            )
             for i, value in enumerate(counts)
         ]
 
-        return cls(bins, observations, mean, std)
+        return cls(bins, observations, float(mean), float(std))
 
 
 @dataclass
